mtpd: l2tp: Fix endianness issues in log prints
Tunnel and session variables are obtained from network, thus are stored
in network endianness. Although this fact is mentioned in this comment:

    // To avoid unnecessary endianness conversions, tunnels, sessions,
    // attributes, and values are all accessed in network order

those variables are not being converted to host endianness when printing.

So convert them to host endianness when printing. Also change the type of
those variables to __be16 so we can distinguish them easier.

Change-Id: If4a147929fbc79ada51ce5c4f6a1e929d76267aa
Signed-off-by: Sam Protsenko <semen.protsenko@linaro.org>