mtpd: Remove obsolete way of passing args to pppd
This Android-specific way of passing args to pppd was added in 4a721df1
to avoid exposing the L2TP secret in /proc/$$/cmdline (b/2325838). But
then 41ab620e removed the secret from the command line altogether. So we
no longer need this hack.

This reverts 4a721df1 ("Use environment variables to pass information to
pppd.").

Change-Id: I0bbe7424c94bdcfb45c8198fa49f8b5c33bf9db6
Signed-off-by: Sam Protsenko <semen.protsenko@linaro.org>